Allow session creation using specific creation time.

+++ marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/ByteBufferInputStream.java
@@ -45,7 +45,7 @@ public class ByteBufferInputStream extends InputStream {
 
 	@Override
 	public long skip(long bytes) {
-		int offset = (int) Math.min((long) this.buffer.remaining(), bytes);
+		int offset = (int) Math.min(this.buffer.remaining(), bytes);
 		this.buffer.position(this.buffer.position() + offset);
 		return offset;
 	}
